+++ lib/src/errors/status_errors.dart
@@ -1,19 +1,28 @@
 import 'dart:io';
 
-class StatusCode {
-  static e404(HttpRequest req) {
+abstract class StatusCodes {
+  e404(HttpRequest req);
+  e405(HttpRequest req);
+  e500(HttpRequest req);
+}
+
+class TextStatusCodes extends StatusCodes {
+  @override
+  e404(HttpRequest req) {
     req.response.statusCode = 404;
     req.response.write('Page not found.');
     req.response.close();
   }
 
-  static e405(HttpRequest req) {
+  @override
+  e405(HttpRequest req) {
     req.response.statusCode = 405;
     req.response.write('Method not allowed.');
     req.response.close();
   }
 
-  static e500(HttpRequest req) {
+  @override
+  e500(HttpRequest req) {
     req.response.statusCode = 500;
     req.response.write('Server error.');
     req.response.close();